move makeRuntime down to module
