+++ src/api/run_cucumber.ts
@@ -6,8 +6,8 @@ import { resolvePaths } from '../paths'
 import { SupportCodeLibrary } from '../support_code_library_builder/types'
 import { version } from '../version'
 import { IFilterablePickle } from '../filter'
+import { makeRuntime } from '../runtime'
 import { IRunOptions, IRunEnvironment, IRunResult } from './types'
-import { makeRuntime } from './runtime'
 import { initializeFormatters } from './formatters'
 import { getSupportCodeLibrary } from './support'
 import { mergeEnvironment } from './environment'
