+++ test/formatter_helpers.ts
@@ -3,7 +3,7 @@ import { PassThrough } from 'node:stream'
 import { promisify } from 'node:util'
 import { IdGenerator } from '@cucumber/messages'
 import * as messages from '@cucumber/messages'
-import { RuntimeOptions } from '../src/runtime'
+import { makeRuntime, RuntimeOptions } from '../src/runtime'
 import { EventDataCollector } from '../src/formatter/helpers'
 import FormatterBuilder from '../src/formatter/builder'
 import { SupportCodeLibrary } from '../src/support_code_library_builder/types'
@@ -12,7 +12,6 @@ import { doesNotHaveValue } from '../src/value_checker'
 import { emitSupportCodeMessages } from '../src/cli/helpers'
 import { FormatOptions } from '../src/formatter'
 import { SourcedPickle } from '../src/assemble'
-import { makeRuntime } from '../src/api/runtime'
 import { IRunEnvironment } from '../src/api'
 import { generatePickles } from './gherkin_helpers'
 import { buildOptions, buildSupportCodeLibrary } from './runtime_helpers'
